Add source for mitf action
The logic of the action is mostly split into two parts:

1. Checking if the usage message was requested.
2. If not, parsing the date and running a MIT command.

The action, as of this patch, requires the TODO text to placed as the
last argument, surrounded by quotation marks. Would one not quote the
TODO, only the first word will be used. This behaviour was chosen due to
my habit of always surrounding TODOs with quotation marks when I use
todo.sh. A potential improvement would be to allow for omitting these
quotation marks. For my own usage, this won't be necessary, however.