Extract methods with intent revealing names
While it's fairly obvious that 'd', 'w' and 'm' refer to days, weeks and
months, the method extractions in this patch makes the intent a bit more
clear.

Additionally, the month incrementing with `>>` might not be something
that every reader is familiar with, so the method extraction helps with
that as well.